--- test/browser/feedback-pages-helper.test.js
@@ -1,4 +1,0 @@
-/* eslint-env jest */
-'use strict'
-
-require('../helpers/feedback-pages-helper.test')
