+++ test/helpers/feedback-pages.test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest */
 'use strict'
 
-const FeedbackPagesHelper = require('../../src/helpers/feedback-pages-helper')
+const FeedbackPages = require('../../src/helpers/feedback-pages')
 
 describe('FeedbackPagesHelper', () => {
   const baseUrl = 'https://www.example.org/feedback'
@@ -9,7 +9,7 @@ describe('FeedbackPagesHelper', () => {
 
   describe('.buildUrl', () => {
     describe('without custom translations', () => {
-      const subject = FeedbackPagesHelper.buildUrl(baseUrl, path)
+      const subject = FeedbackPages.buildUrl(baseUrl, path)
 
       it('creates URL to feedback pages with path', () => {
         expect(subject).toEqual(`${baseUrl}/${path.join('/')}`)
@@ -26,7 +26,7 @@ describe('FeedbackPagesHelper', () => {
         }
       }
 
-      const subject = FeedbackPagesHelper.buildUrl(baseUrl, path, translations)
+      const subject = FeedbackPages.buildUrl(baseUrl, path, translations)
 
       it('creates URL to feedback pages with path and custom translations', () => {
         expect(subject).toEqual(`${baseUrl}/${path.join('/')}?translations%5Bfull_name%5D=John%20Doe&translations%5Bitem_type%5D%5Bde%5D=Gruppe&translations%5Bitem_type%5D%5Ben%5D=Group`)
@@ -36,7 +36,7 @@ describe('FeedbackPagesHelper', () => {
 
   describe('.parseTranslations', () => {
     const queryString = '?translations%5Bfull_name%5D=John%20Doe&translations%5Bitem_type%5D%5Bde%5D=Gruppe&translations%5Bitem_type%5D%5Ben%5D=Group`'
-    const subject = FeedbackPagesHelper.parseTranslations(queryString, 'de')
+    const subject = FeedbackPages.parseTranslations(queryString, 'de')
 
     it('return object with translations', () => {
       expect(subject).toEqual({
