+++ test/node/feedback-pages.test.js
@@ -1,0 +1,4 @@
+/* eslint-env jest */
+'use strict'
+
+require('../helpers/feedback-pages.test')
